Introduce protocol::bob::behaviour to mimic structure from alice module
